+++ app/Http/Controllers/Dashboard/FAQController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Dashboard;
 use App\Http\Controllers\Controller;
 use App\Models\FAQ;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use RealRashid\SweetAlert\Facades\Alert;
 
 class FAQController extends Controller
@@ -12,7 +13,11 @@ class FAQController extends Controller
     public function index()
     {
         try {
-
+            $user = Auth::guard('admin')->user();
+            if (!$user || !$user->hasPermissionByRole('view faq')) {
+                Alert::toast('You dont have access to this page.','error');
+                return redirect()->back();
+            }
             $faq = FAQ::paginate(10);
             return view('faq.index', compact('faq'));
         } catch (\Exception $e) {
@@ -25,7 +30,11 @@ class FAQController extends Controller
     public function create()
     {
         try {
-
+            $user = Auth::guard('admin')->user();
+            if (!$user || !$user->hasPermissionByRole('add faq')) {
+                Alert::toast('You dont have access to this page.','error');
+                return redirect()->back();
+            }
             return view('faq.create');
         } catch (\Exception $e) {
             // Handle exceptions or errors
@@ -36,8 +45,12 @@ class FAQController extends Controller
 
     public function store(Request $request)
     {
-        // try {
-
+        try {
+            $user = Auth::guard('admin')->user();
+            if (!$user || !$user->hasPermissionByRole('add faq')) {
+                Alert::toast('You dont have access to this page.','error');
+                return redirect()->back();
+            }
             if (!$request->isMethod('post')) {
                 // Handle the error - Method not allowed
                 Alert::toast('Method not allowed', 'error');
@@ -57,20 +70,24 @@ class FAQController extends Controller
 
             Alert::toast('faq has been saved successfully!', 'success');
             return redirect()->route('faqs');
-        // } catch (\Illuminate\Validation\ValidationException $e) {
-        //     // Laravel's built-in validation exception
-        //     return redirect()->back()->withErrors($e->validator->errors())->withInput();
-        // } catch (\Exception $e) {
-        //     // Handle exceptions or errors
-        //     Alert::toast('something is wrong!!', 'error');
-        //     return redirect()->back();
-        // }
+        } catch (\Illuminate\Validation\ValidationException $e) {
+            // Laravel's built-in validation exception
+            return redirect()->back()->withErrors($e->validator->errors())->withInput();
+        } catch (\Exception $e) {
+            // Handle exceptions or errors
+            Alert::toast('something is wrong!!', 'error');
+            return redirect()->back();
+        }
     }
 
     public function edit($id)
     {
         try {
-
+            $user = Auth::guard('admin')->user();
+            if (!$user || !$user->hasPermissionByRole('edit faq')) {
+                Alert::toast('You dont have access to this page.','error');
+                return redirect()->back();
+            }
             $faq = FAQ::find($id);
             return view('faq.edit', compact('faq'));
 
@@ -84,7 +101,11 @@ class FAQController extends Controller
     public function update(Request $request)
     {
         try {
-
+            $user = Auth::guard('admin')->user();
+            if (!$user || !$user->hasPermissionByRole('edit faq')) {
+                Alert::toast('You dont have access to this page.','error');
+                return redirect()->back();
+            }
             if (!$request->isMethod('put')) {
                 // Handle the error - Method not allowed
                 Alert::toast('Method not allowed', 'error');
@@ -111,6 +132,11 @@ class FAQController extends Controller
     public function delete($id)
     {
         try {
+            $user = Auth::guard('admin')->user();
+            if (!$user || !$user->hasPermissionByRole('delete faq')) {
+                Alert::toast('You dont have access to this page.','error');
+                return redirect()->back();
+            }
             $faq = FAQ::find($id);
             $faq->delete();
             Alert::toast('faq has been deleted!', 'error');
@@ -125,6 +151,11 @@ class FAQController extends Controller
     public function active($id)
     {
         try {
+            $user = Auth::guard('admin')->user();
+            if (!$user || !$user->hasPermissionByRole('edit faq')) {
+                Alert::toast('You dont have access to this page.','error');
+                return redirect()->back();
+            }
             $faq = FAQ::find($id);
             $faq->status = 1;
             $faq->save();
@@ -141,7 +172,11 @@ class FAQController extends Controller
     public function inactive($id)
     {
         try {
-
+            $user = Auth::guard('admin')->user();
+            if (!$user || !$user->hasPermissionByRole('edit faq')) {
+                Alert::toast('You dont have access to this page.','error');
+                return redirect()->back();
+            }
             $faq = FAQ::find($id);
             $faq->status = 0;
             $faq->save();
